Unit Slider
Unit Slider hooked up to planet trade menu code

+++ Assets/Scripts/Commodity.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "commodity", menuName = "Commodity")]
 public class Commodity : ScriptableObject {
 
-    public int worth;
+    public float worth;
 }
